renames files to be consistant with the rest of the project
