+++ components/search-result/SearchResult.module.css
